QQControls; Qt.labs.handlers; better selection behavior; targeting Qt 5.10
Qt.labs.handlers will be the 5.10 release of the Pointer Handlers concept,
as tech preview.  This works with that.  There is no MouseHandler ready for
release, so far, so we have to abuse DragHandler to enable dragging to
select text.

Putting the Selection at z: 1 puts it on top of the text.

Coordinate mapping was confusingly implemented, and wrong too.
So now the selection actually follows the cursor properly, even
if there is a lot of scrollback text.